Move entities to data, allow null BalanceToken
