+++ src/common/schemas/json-schema.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import Ajv, { JSONSchemaType, ValidateFunction } from 'ajv';
+import Ajv, { JSONSchemaType, Schema, ValidateFunction } from 'ajv';
 
 @Injectable()
 export class JsonSchemaService {
@@ -7,14 +7,14 @@ export class JsonSchemaService {
 
   constructor() {
     // coerceTypes param shouldn't be necessary when serialization is implemented.
-    this.ajv = new Ajv({ coerceTypes: true });
+    this.ajv = new Ajv({ coerceTypes: true, allowUnionTypes: true });
   }
 
   addSchema<T>(schema: JSONSchemaType<T>, name: string): void {
     this.ajv.addSchema(schema, name);
   }
 
-  compile<T>(schema: JSONSchemaType<T>): ValidateFunction {
+  compile<T>(schema: Schema | JSONSchemaType<T>): ValidateFunction {
     return this.ajv.compile(schema);
   }
 }
