+++ src/datasources/exchange-api/entities/fiat-codes-result.entity.ts
@@ -1,4 +1,5 @@
 export interface FiatCodesExchangeResult {
   success: boolean;
+  // TODO: use Record<string, string>, define a compatible JSONSchema
   symbols: object;
 }
